Adicionando validaçao para Editar e Deletar, somente Se existir

+++ app/Http/Controllers/ContactController.php
@@ -12,17 +12,24 @@ class ContactController extends Controller
      */
     public function index()
     {
-          // Busca todos os contatos
-          $contacts = Contact::orderBy('id', 'DESC')->paginate(5);
-          // Retorna a view index com os contatos
-          return view('contacts.index', ['contacts' => $contacts]);
+        // Busca todos os contatos
+        $contacts = Contact::orderBy('id', 'DESC')->paginate(5);
+        // Retorna a view index com os contatos
+        return view('contacts.index', ['contacts' => $contacts]);
     }
 
     public function destroy(Contact $contact)
     {
-        $contact->delete();
+        // Verificar se o contato existe
+        if ($contact->exists()) {
+            // Deletar o contato
+            $contact->delete();
 
-        // Redirecionar de volta para a página da Lista
-        return redirect()->route('contacts.index')->with('success', 'Contact deleted successfully');
+            // Redirecionar de volta para a página da Lista
+            return redirect()->route('contacts.index')->with('success', 'Contact deleted successfully');
+        } else {
+            // Se o contato não existe, redirecionar com uma mensagem de erro
+            return redirect()->route('contacts.index')->with('error', 'Contact not found.');
+        }
     }
 }
